network: fix cross dependencies
